# Fix crash during lua script defrag