@oranagra 2025-09-03
not with newly generated AOF files, in which scripts propagate effects rather than scripts.
but you're right, either change the check to verify you're not in script, or, i think it's better, move that LRU list move to the beginning of the script rather than it's end.